+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added Notion integration for exporting articles directly to Notion pages. (#625)
+
 ## [1.3.18] - 2026-01-29
 
 ### Refactored